Cleanup processing listener
